+++ packages/frontend/src/components/MkAudioVisualizer.vue
